# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API мини-приложения реверс-найма в MAX.\nАвторизация: POST /api/v1/auth c initData из MAX Bridge,\nдалее заголовок Authorization: Bearer \u003ctoken\u003e (JWT, 7 дней).",
        "title": "MAX Mini App — Reverse Hiring API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Текущий пользователь",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.MeResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/me/role": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Выбор роли (+ согласие на обработку ПДн)",
                "parameters": [
                    {
                        "description": "роль и согласие",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.RoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.MeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/my/referrals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Мои рефералы",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.referralResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/vacancies/map": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vacancy"
                ],
                "summary": "Вакансии на карте",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/max-miniapp_backend_internal_dto.MapVacancy"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vacancies/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vacancy"
                ],
                "summary": "Обновить вакансию",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID вакансии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "изменяемые поля",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.VacancyUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.Vacancy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_handler.referralResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/max-miniapp_backend_internal_dto.ReferralItem"
                    }
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.MapVacancy": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "company_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "salary_max": {
                    "type": "integer"
                },
                "salary_min": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.MeResponse": {
            "type": "object",
            "properties": {
                "personal_data_accepted_at": {
                    "type": "string"
                },
                "referral_code": {
                    "type": "string",
                    "example": "ref_42"
                },
                "user": {
                    "$ref": "#/definitions/max-miniapp_backend_internal_dto.User"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.ReferralItem": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "joined_at": {
                    "type": "string"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.RoleRequest": {
            "type": "object",
            "properties": {
                "acceptPersonalData": {
                    "type": "boolean",
                    "example": true
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "candidate",
                        "recruiter"
                    ],
                    "example": "candidate"
                }
            }
        },
        "max-miniapp_backend_internal_dto.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "example": "Ivan"
                },
                "id": {
                    "type": "integer",
                    "example": 42
                },
                "language_code": {
                    "type": "string",
                    "example": "ru"
                },
                "last_name": {
                    "type": "string",
                    "example": "Petrov"
                },
                "photo_url": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "example": "candidate"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "example": "ivan"
                }
            }
        },
        "max-miniapp_backend_internal_dto.Vacancy": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Москва"
                },
                "company_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "Платёжный сервис"
                },
                "employment_type": {
                    "type": "string",
                    "enum": [
                        "full_time",
                        "part_time",
                        "contract",
                        "internship"
                    ],
                    "example": "full_time"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "min_experience_months": {
                    "type": "integer",
                    "example": 12
                },
                "required_skills": {
                    "type": "string",
                    "example": "go, postgres"
                },
                "response_ttl_hours": {
                    "type": "integer",
                    "example": 48
                },
                "salary_max": {
                    "type": "integer",
                    "example": 280000
                },
                "salary_min": {
                    "type": "integer",
                    "example": 180000
                },
                "title": {
                    "type": "string",
                    "example": "Go-разработчик"
                },
                "updated_at": {
                    "type": "string"
                },
                "work_format": {
                    "type": "string",
                    "enum": [
                        "onsite",
                        "hybrid",
                        "remote"
                    ],
                    "example": "hybrid"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.VacancyUpdate": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "min_experience_months": {
                    "type": "integer"
                },
                "response_ttl_hours": {
                    "type": "integer",
                    "example": 72
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Значение: \"Bearer \u003cjwt\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}